+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+### Fixed
+* Unions with custom metaclasses supported
 
 ## [1.11](https://pypi.org/project/multimethod/1.11/) - 2024-01-28
 ### Changed