+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## Unreleased
 ### Fixed
 * Unions with custom metaclasses supported
+* Concrete protocol classes supported
 
 ## [1.11](https://pypi.org/project/multimethod/1.11/) - 2024-01-28
 ### Changed