+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+
+## [1.11.1](https://pypi.org/project/multimethod/1.11.1/) - 2024-02-19
 ### Fixed
 * Unions with custom metaclasses supported
 * Concrete protocol classes supported